Don't set Cache-Control to no-cache if cache enabled is null
If the user has not asked GeoServer to set the cache-control response header then we should not set the header at all as it might be overwriting an existing header.

This commit is a possible update which might achieve this result but I am not able to build yet so this is untested.